# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HST API Documentation",
        "description": "Application API services documentation page",
        "contact": {
            "email": "volkangurgor@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/majors": {
            "get": {
                "tags": [
                    "Majors"
                ],
                "summary": "Get all majors",
                "operationId": "a6fb87f871993e5c7f2dc4de65d1a2ac",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Majors"
                ],
                "summary": "Create a new major",
                "operationId": "994bd62803e7276905c33d2a5fcc8f27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/majors/{id}": {
            "get": {
                "tags": [
                    "Majors"
                ],
                "summary": "Get a specific major",
                "operationId": "665a720de4f28928fd275478d43995de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Major ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Major not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Majors"
                ],
                "summary": "Update a major",
                "operationId": "aaec71342c68e14604f6d46b5bc8621f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Major ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Major not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Majors"
                ],
                "summary": "Delete a major",
                "operationId": "d9c302edf224d0fdaa701ac9268f8257",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Major ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Major not found"
                    }
                }
            }
        },
...
        "/api/lesson-slots": {
            "get": {
                "tags": [
                    "Lesson Slots"
                ],
                "summary": "Get all lesson slots",
                "operationId": "964bb122819810b19664f9910c4306cf",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Lesson Slots"
                ],
                "summary": "Create a new lesson slot",
                "operationId": "ccf900a24482f3ce1fd77fd6fa22b7dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "campus_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "day": {
                                        "type": "string",
                                        "example": "Monday"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "09:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "10:30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/lesson-slots/{id}": {
            "get": {
                "tags": [
                    "Lesson Slots"
                ],
                "summary": "Get a specific lesson slot",
                "operationId": "dc42997189f120cc0891b10427e2a3a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson Slot ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Lesson slot not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Lesson Slots"
                ],
                "summary": "Update a lesson slot",
                "operationId": "ece2e7f48735c033ebb833ce0316b400",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson Slot ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "campus_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "day": {
                                        "type": "string",
                                        "example": "Tuesday"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "10:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "11:30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Lesson slot not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lesson Slots"
                ],
                "summary": "Delete a lesson slot",
                "operationId": "8d5a437e4e9c09e7edc3f40bc0a181d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson Slot ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Lesson slot not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "name": "Authorization",
                "description": "Bearer token to access these API endpoints",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}